Teams system 1

+++ src/main/java/net/nxtresources/managers/LocationManager.java
@@ -44,9 +44,9 @@ public class LocationManager {
         World world = Bukkit.getWorld(parts[0]);
         if (world == null) return null;
 
-        int x = Integer.parseInt(parts[1]);
-        int y = Integer.parseInt(parts[2]);
-        int z = Integer.parseInt(parts[3]);
+        double x = Double.parseDouble(parts[1]);
+        double y = Double.parseDouble(parts[2]);
+        double z = Double.parseDouble(parts[3]);
         float yaw = Float.parseFloat(parts[4]);
         float pitch = Float.parseFloat(parts[5]);
